% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dir_setup}
\alias{dir_setup}
\title{Set up processing folders}
\usage{
dir_setup()
}
\value{
creates data and dataout folder
}
\description{
Set up processing folders
}
\examples{
\dontrun{
#standard
  dir_setup()
}
}
