% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_id.R
\docType{data}
\name{disagg_map_id}
\alias{disagg_map_id}
\title{USAID Disaggregate Mapping File Google Sheet ID}
\format{
An object of class \code{character} of length 1.
}
\usage{
disagg_map_id
}
\value{
google sheet id
}
\description{
USAID Disaggregate Mapping File Google Sheet ID
}
\keyword{internal}
