% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Get period metadata}
\usage{
get_meta(period)
}
\arguments{
\item{period}{Input reporting period}
}
\description{
Returns current fiscal year, qtr, and import file period to store as locals in env
}
\examples{
\dontrun{
get_meta("FY23Q4")
}
}
