% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_mech_data.R
\name{grab_mech_data}
\alias{grab_mech_data}
\title{Grab site-level mechanism mechdata from DATIM/recent MSD}
\usage{
grab_mech_data(mech_df, msd_df, extra_mechs = FALSE)
}
\arguments{
\item{mech_df}{Calls on mechanism meta-data that is pulled by \code{pull_mech_uid()}}

\item{msd_df}{Calls on latest MSD data pulled by \code{gophr::read_msd()}}

\item{extra_mechs}{If TRUE, function will read in extra mechanism data saved in reference file folder
(default = FALSE)}
}
\value{

}
\description{
Grab site-level mechanism mechdata from DATIM/recent MSD
}
\examples{
 \dontrun{
   grab_mech_data(df = mechs, extra_mechs = TRUE)
}

}
