% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kp_names}
\alias{kp_names}
\title{Standard names for KP tabs in NDOH}
\format{
A character vector with length 2
}
\usage{
kp_names
}
\description{
A character vector of columns in each KP tab of NDOH (KP_Location, KP_type)
}
\keyword{datasets}
