% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_disaggs.R
\name{map_disaggs}
\alias{map_disaggs}
\title{Mapping DATIM UIDS to NDOH tab structure}
\usage{
map_disaggs(df, ind_sel = "All", disaggregate = "All", all_indic = TRUE)
}
\arguments{
\item{df}{NDOH dataset to map DATIM hierachy into}

\item{ind_sel}{Selected indicators to map - default is "All" indicators,
which removes PrEP_CT and TX_RTT as they must be mapped separately ("All", "PrEP_CT", "TX_RTT")}

\item{disaggregate}{Selected disaggregate group to map - default is for all indicators (minus PrEP_CT & TX_RTT)
i.e. PrEP_CT Disaggs: (1) "Age/Sex" (2) "TestResult", TX_RTT Disaggs: (1) "Age/Sex/HIVStatus" (2) "ARTNoContactReasonIIT}

\item{all_indic}{TRUE if data for all indicators (minus PrEP_CT and TX_RTT) are being mapped, FALSE if not}
}
\value{

}
\description{
Mapping DATIM UIDS to NDOH tab structure
}
\examples{
 \dontrun{
   map_disaggs(df = ndoh_clean, ind_sel = "PrEP_CT", disaggregate ="Age/Sex", all_indic = FALSE)
   map_disaggs(df = ndoh_clean, all_indic = TRUE)
}
}
