% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_id.R
\docType{data}
\name{mfl_fy23_id}
\alias{mfl_fy23_id}
\title{FY23 Master Facility List Google Sheet ID}
\format{
An object of class \code{character} of length 1.
}
\usage{
mfl_fy23_id
}
\value{
google sheet id
}
\description{
FY23 Master Facility List Google Sheet ID
}
\keyword{internal}
