% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_id.R
\docType{data}
\name{mfl_new_id}
\alias{mfl_new_id}
\title{Master Facility List Google Sheet ID}
\format{
An object of class \code{character} of length 1.
}
\usage{
mfl_new_id
}
\value{
google sheet id
}
\description{
Master Facility List Google Sheet ID
}
\keyword{internal}
