% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_arvdisp}
\alias{names_arvdisp}
\title{Standard names for ARVDISP tab in NDOH}
\format{
An object of class \code{character} of length 9.
}
\usage{
names_arvdisp
}
\description{
A character vector of columns in ARVDISP tab of NDOH (TIER)
}
\keyword{datasets}
