% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_hts_tst}
\alias{names_hts_tst}
\title{Standard names for HTS_TST tab in NDOH}
\format{
A character vector with length 10
}
\usage{
names_hts_tst
}
\description{
A character vector of columns in HTS_TST tab of NDOH (non-TIER)
}
\keyword{datasets}
