% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_hts_tst_kp}
\alias{names_hts_tst_kp}
\title{Standard names for HTS_TST KP tab in NDOH}
\format{
A character vector with length 12
}
\usage{
names_hts_tst_kp
}
\description{
A character vector of columns in HTS_TST KP tab of NDOH (non-TIER)
}
\keyword{datasets}
