% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_pmtct_art}
\alias{names_pmtct_art}
\title{Standard names for PMTCT_ART tab in NDOH}
\format{
A character vector with length 8
}
\usage{
names_pmtct_art
}
\description{
A character vector of columns in PMTCT_ART tab of NDOH (non-TIER)
}
\keyword{datasets}
