% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_pmtct_eid}
\alias{names_pmtct_eid}
\title{Standard names for PMTCT_EID tab in NDOH}
\format{
A character vector with length 7
}
\usage{
names_pmtct_eid
}
\description{
A character vector of columns in PMTCT_EID tab of NDOH (non-TIER)
}
\keyword{datasets}
