% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_pmtct_hei_pos}
\alias{names_pmtct_hei_pos}
\title{Standard names for PMTCT_HEI_POS tab in NDOH}
\format{
A character vector with length 7
}
\usage{
names_pmtct_hei_pos
}
\description{
A character vector of columns in PMTCT_HEI_POS tab of NDOH (non-TIER)
}
\keyword{datasets}
