% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_prep_ct}
\alias{names_prep_ct}
\title{Standard names for PrEP_CT tab in NDOH}
\format{
A character vector with length 9
}
\usage{
names_prep_ct
}
\description{
A character vector of columns in PrEP_CT tab of NDOH (TIER)
}
\keyword{datasets}
