% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_prep_ct_kp}
\alias{names_prep_ct_kp}
\title{Standard names for PrEP_CT KP tab in NDOH}
\format{
A character vector with length 11
}
\usage{
names_prep_ct_kp
}
\description{
A character vector of columns in PrEP_CT KP tab of NDOH (TIER)
}
\keyword{datasets}
