% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tb_art}
\alias{names_tb_art}
\title{Standard names for TB_ART tab in NDOH}
\format{
A character vector with length 9
}
\usage{
names_tb_art
}
\description{
A character vector of columns in TB_ART tab of NDOH (TIER)
}
\keyword{datasets}
