% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tb_prev}
\alias{names_tb_prev}
\title{Standard names for TB_PREV tab in NDOH}
\format{
A character vector with length 9
}
\usage{
names_tb_prev
}
\description{
A character vector of columns in TB_PREV tab of NDOH (TIER)
}
\keyword{datasets}
