% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tb_stat_d}
\alias{names_tb_stat_d}
\title{Standard names for TB_STAT_D tab in NDOH}
\format{
A character vector with length 8
}
\usage{
names_tb_stat_d
}
\description{
A character vector of columns in TB_STAT_D tab of NDOH (TIER)
}
\keyword{datasets}
