% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tb_stat_n}
\alias{names_tb_stat_n}
\title{Standard names for TB_STAT_N tab in NDOH}
\format{
A character vector with length 9
}
\usage{
names_tb_stat_n
}
\description{
A character vector of columns in TB_STAT_N tab of NDOH (TIER)
}
\keyword{datasets}
