% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_curr}
\alias{names_tx_curr}
\title{Standard names for TX_CURR tab in NDOH}
\format{
A character vector with length 8
}
\usage{
names_tx_curr
}
\description{
A character vector of columns in TX_CURR tab of NDOH (TIER)
}
\keyword{datasets}
