% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_ml}
\alias{names_tx_ml}
\title{Standard names for TX_ML tab in NDOH}
\format{
A character vector with length 9
}
\usage{
names_tx_ml
}
\description{
A character vector of columns in TX_ML tab of NDOH (TIER)
}
\keyword{datasets}
