% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_ml_kp}
\alias{names_tx_ml_kp}
\title{Standard names for TX_ML KP tab in NDOH}
\format{
A character vector with length 11
}
\usage{
names_tx_ml_kp
}
\description{
A character vector of columns in TX_ML KP tab of NDOH (TIER)
}
\keyword{datasets}
