% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_new}
\alias{names_tx_new}
\title{Standard names for TX_NEW tab in NDOH}
\format{
A character vector with length 8
}
\usage{
names_tx_new
}
\description{
A character vector of columns in TX_NEW tab of NDOH (TIER)
}
\keyword{datasets}
