% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_new_kp}
\alias{names_tx_new_kp}
\title{Standard names for TX_NEW KP tab in NDOH}
\format{
A character vector with length 10
}
\usage{
names_tx_new_kp
}
\description{
A character vector of columns in TX_NEW KP tab of NDOH (TIER)
}
\keyword{datasets}
