% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_pvls}
\alias{names_tx_pvls}
\title{Standard names for TX_PVLS tab in NDOH}
\format{
A character vector with length 8
}
\usage{
names_tx_pvls
}
\description{
A character vector of columns in TX_PVLS tab of NDOH (TIER)
}
\keyword{datasets}
