% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_pvls_n}
\alias{names_tx_pvls_n}
\title{Standard names for TX PVLS N tab in NDOH}
\format{
A character vector with length 8
}
\usage{
names_tx_pvls_n
}
\description{
A character vector of columns in TX PVLS N tab of NDOH with CD4 (TIER)
}
\keyword{datasets}
