% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_pvls_n_kp}
\alias{names_tx_pvls_n_kp}
\title{Standard names for TX PVLS N KP tab in NDOH}
\format{
A character vector with length 8
}
\usage{
names_tx_pvls_n_kp
}
\description{
A character vector of columns in TX PVLS N KP tab of NDOH with CD4 (TIER)
}
\keyword{datasets}
