% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_rtt}
\alias{names_tx_rtt}
\title{Standard names for TX_RTT tab in NDOH}
\format{
A character vector with length 9
}
\usage{
names_tx_rtt
}
\description{
A character vector of columns in TX_RTT tab of NDOH (TIER)
}
\keyword{datasets}
