% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_tb_d}
\alias{names_tx_tb_d}
\title{Standard names for TX_TB_D tab in NDOH}
\format{
A character vector with length 10
}
\usage{
names_tx_tb_d
}
\description{
A character vector of columns in TX_TB_D tab of NDOH (TIER)
}
\keyword{datasets}
