% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_tb_d_pos}
\alias{names_tx_tb_d_pos}
\title{Standard names for TX_TB_D_POs tab in NDOH}
\format{
An object of class \code{character} of length 6.
}
\usage{
names_tx_tb_d_pos
}
\description{
A character vector of columns in TX_TB_D_POs tab of NDOH (TIER)
}
\keyword{datasets}
