% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_tb_d_testtype}
\alias{names_tx_tb_d_testtype}
\title{Standard names for TX_TB_D TestType tab in NDOH}
\format{
A character vector with length 8
}
\usage{
names_tx_tb_d_testtype
}
\description{
A character vector of columns in X_TB_D TestType tab tab of NDOH (TIER)
}
\keyword{datasets}
