% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_tx_tb_n}
\alias{names_tx_tb_n}
\title{Standard names for TX_TB_N tab in NDOH}
\format{
A character vector with length 9
}
\usage{
names_tx_tb_n
}
\description{
A character vector of columns in TX_TB_N tab of NDOH (TIER)
}
\keyword{datasets}
