% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_disaggs.R
\name{ndoh_post_processing}
\alias{ndoh_post_processing}
\title{NDOH Post Processing Function}
\usage{
ndoh_post_processing(kp = FALSE, export_type)
}
\arguments{
\item{kp}{parameter to map and process KP tabs if KP = TRUE (default KP = FALSE)}

\item{export_type}{parameter to extract data in specific format (1) DATIM import file (2) Validation file for internal checks}
}
\value{

}
\description{
This function maps in DATIM elements using \code{map_disaggs()}, binds all the indicator levels together &
exports processed data in import file or validation file format
}
\examples{
 \dontrun{
   ndoh_post_processing(kp = FALSE, export_type = "Import")
}
}
