% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndoh_processing.R
\name{ndoh_wrapper}
\alias{ndoh_wrapper}
\title{Wrapper function to process NDOH data}
\usage{
ndoh_wrapper(
  filepath = ndoh_filepath,
  qtr = curr_qtr,
  kp = FALSE,
  export_type = "Import"
)
}
\arguments{
\item{filepath}{NDOH file path}

\item{qtr}{current quarter for NDOH processing}

\item{kp}{processes KP tabs in NDOH file if TRUE (default is FALSE) - need to run both and bind together}

\item{export_type}{parameter to extract data in specific format (1) DATIM import file (2) Validation file for internal checks}
}
\value{

}
\description{
This function wraps together functions to import the NDOH file and append all indicator tabs
(\code{import_ndoh()}), checks for facility alignment between NDOH and the MFL (\code{validate_ndoh()}), tidies the
dataframe and cleans up indicator names (\code{tidy_ndoh()}), and finally maps in the dataElement and categoryOptionCombo
information using \code{map_disaggs()} and \code{ndoh_post_processing()}
}
\examples{
\dontrun{

  df <- ndoh_wrapper(ndoh_filepath, qtr = "Q3", kp = TRUE, export_type = "Validation")
}
}
