% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_groups.R
\docType{data}
\name{non_tier_qtr}
\alias{non_tier_qtr}
\title{Quarterly non-TIER indicators reported in NDOH file}
\format{
An object of class \code{character} of length 5.
}
\usage{
non_tier_qtr
}
\description{
This is the list of non-TIER indicators that are reported in NDOH
}
\keyword{datasets}
