% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndoh_processing.R
\name{partner_import}
\alias{partner_import}
\title{Split import file into partner import files for DQRT}
\usage{
partner_import(df, mech_code)
}
\arguments{
\item{df}{TIER import file from \code{ndoh_processing()}}

\item{mech_code}{partner mech code (numeric)}
}
\value{

}
\description{
Split import file into partner import files for DQRT
}
