% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_mech_uid.R
\name{pull_mech_uid}
\alias{pull_mech_uid}
\title{Pull Partner/Mechanism Info from DATIM}
\usage{
pull_mech_uid(usaid_only = TRUE, ou_sel = NULL, folderpath_output = NULL)
}
\arguments{
\item{usaid_only}{specify if only USAID mechansism should be returned, default = TRUE}

\item{ou_sel}{option to specify an operating unit, default = NULL}

\item{folderpath_output}{provide the full path to the folder for saving}
}
\description{
Pull Partner/Mechanism Info from DATIM
}
\examples{
\dontrun{
#pull mechanism/partner information
df <- pull_mech_uid(ou_sel = "South Africa") }
}
