% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_names}
\alias{standard_names}
\title{Standard names for each indicator tab in NDOH}
\format{
A character vector with length 5
}
\usage{
standard_names
}
\description{
A character vector of standard column names for each tab (Province, District, Subdistrict, Facility, Code)
}
\keyword{datasets}
