% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ndoh.R
\name{tidy_map_arvdisp}
\alias{tidy_map_arvdisp}
\title{Tidy and map disaggs for ARVDISP}
\usage{
tidy_map_arvdisp(df)
}
\arguments{
\item{df}{dataframe from import_arvdisp()}
}
\description{
Tidy and map disaggs for ARVDISP
}
\examples{
 \dontrun{
   import_arvdisp(ndoh_filepath) \%>\% tidy_map_arvdisp
}

}
