% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ndoh.R
\name{tidy_ndoh}
\alias{tidy_ndoh}
\title{Tidy / munge NDOH import}
\usage{
tidy_ndoh(df, kp = FALSE)
}
\arguments{
\item{df}{imported dataframe to pass in}

\item{kp}{parameter to tidy KP tabs if KP = TRUE (default KP = FALSE)}
}
\value{

}
\description{
Tidy / munge NDOH import
}
\examples{
 \dontrun{
   tidy_ndoh(ndoh_all, kp = FALSE)
}

}
