% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_groups.R
\docType{data}
\name{tier_ann}
\alias{tier_ann}
\title{Annual TIER indicators (use in Q4 only)}
\format{
An object of class \code{character} of length 15.
}
\usage{
tier_ann
}
\description{
This is the list of indicators to process from NDOH in Q4 (note: process SC_ARVDISP separately)
}
\keyword{datasets}
