% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_groups.R
\docType{data}
\name{tier_kp_indic}
\alias{tier_kp_indic}
\title{Quarterly TIER indicators for KP tabs (use in each quarter)}
\format{
An object of class \code{character} of length 7.
}
\usage{
tier_kp_indic
}
\description{
This is the list of KP tab indicators to process from NDOH in all quarters
}
\keyword{datasets}
