% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_groups.R
\docType{data}
\name{tier_qtr}
\alias{tier_qtr}
\title{Quarterly TIER indicators (use Q1 and Q3)}
\format{
An object of class \code{character} of length 10.
}
\usage{
tier_qtr
}
\description{
This is the list of indicators to process from NDOH in Q1 and Q3
}
\keyword{datasets}
