% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indic_groups.R
\docType{data}
\name{tier_semi}
\alias{tier_semi}
\title{Semi-Annual TIER indicators (use in Q2 only)}
\format{
An object of class \code{character} of length 14.
}
\usage{
tier_semi
}
\description{
This is the list of indicators to process from NDOH in Q2 (note: process SC_ARVDISP separately)
}
\keyword{datasets}
