% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaid-dsp-district.R
\docType{data}
\name{usaid_dsp_district}
\alias{usaid_dsp_district}
\title{USAID South Africa Districts}
\format{
An object of class \code{character} of length 15.
}
\usage{
usaid_dsp_district
}
\description{
This is the list of PSNUs that USAID South Africa DSPs work in (length = 15)
}
\keyword{datasets}
