% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ndoh.R
\name{validate_ndoh}
\alias{validate_ndoh}
\title{Validate facilities reported in NDOH by MFL}
\usage{
validate_ndoh(df)
}
\value{

}
\description{
Validate facilities reported in NDOH by MFL
}
\examples{
\dontrun{

  validate_ndoh(ndoh_all)
}
}
