% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mfl.R
\name{clean_mfl}
\alias{clean_mfl}
\title{Reshape MFL}
\usage{
clean_mfl(mfl_period = "FY24Q2")
}
\arguments{
\item{mfl_period}{version of MFL}
}
\description{
Reshape MFL
}
\examples{
\dontrun{

  mfl <- clean_mfl()
}
}
