% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndoh_processing.R
\name{ndoh_processing}
\alias{ndoh_processing}
\title{NDOH Processing Function}
\usage{
ndoh_processing(
  filepath = ndoh_filepath,
  qtr = curr_qtr,
  export_type = "Import"
)
}
\arguments{
\item{filepath}{NDOH filepath}

\item{qtr}{current quarter}

\item{export_type}{parameter to extract data in specific format (1) DATIM import file (2) Validation file for internal checks}
}
\value{

}
\description{
This function calls on the wrapper processing function \code{ndoh_wrapper()} to process
both the non-KP and KP tabs in the NDOH file, and bind them all together
}
\examples{
\dontrun{

  df <- ndoh_processing(ndoh_filepath, qtr = "Q3", export_type = "Import")
}
}
