% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{partner_import}
\alias{partner_import}
\title{Split into partner-level import files}
\usage{
partner_import(df, mech_code)
}
\arguments{
\item{df}{final import file dataframe}

\item{mech_code}{partner mech code}
}
\value{

}
\description{
Split into partner-level import files
}
\examples{

}
