% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{partner_import}
\alias{partner_import}
\title{Split into partner-level import files}
\usage{
partner_import(df, mech_code, save = TRUE)
}
\arguments{
\item{df}{final import file dataframe}

\item{mech_code}{partner mech code}

\item{save}{if TRUE, the function will save dataframe to appropriate export folder (default = TRUE)}
}
\description{
Split into partner-level import files
}
