% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_all_the_things}
\alias{read_all_the_things}
\title{Read tabs from NDOH for each indicator and adjust column names}
\usage{
read_all_the_things(path, sheet)
}
\arguments{
\item{path}{file path for NDOH}

\item{sheet}{tab name}
}
\description{
Read tabs from NDOH for each indicator and adjust column names
}
\examples{
\dontrun{

  df <- read_all_the_things(filepath, "TX_CURR")
}
}
